requirements.txt
boto3==1.28.68
python-dotenv==1.0.0